Add C++ implementations of search.
Currently only for DFS and BFS.

DFS is implemented with two signatures: one that takes a Searchable
interface and one that takes two functions: one for testing the goal and
the other for getting the successors. I would prefer to only have one
since the implementations are nearly identical but I haven't found a
good way to pass member functions.

Also, implementations of DFS and BFS are contained in header files to
avoid linker issues; this may be required but I don't particularly like it.